requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymongo
python-multipart
PyMuPDF
python-docx
Pillow
pytesseract
pydantic
streamlit
requests